Intervalometer example
Added some example code that takes an image repeatedly. Uses the
reset() function of the camera library. This prevents the camera from
sending the same image over and over again to the arduino. I also
changed the filename loop so that it can be easily adjusted to take any
number of images. This new loop probably uses more memory for the
String objects, but it's more readable for beginner programmers.